refactor[webpage]: utilize state pattern
Previously the WebPage constructor had side effects, because it already
fetched and analyzed the websites. Now the page has to be analyzed
manually by calling the respective method. To guarantee that no webpage
is analyzed or translated multiple times or even translated prior to
analyzing the state pattern is used. In case of improper use of the
WebPage class an IllegalStateException is thrown.